pyproject.toml
[tool.poetry]
name = "hwaudit"
version = "4.2.0.2"
description = ""
authors = ["uib GmbH <info@uib.de>"]
maintainers = ["uib GmbH <info@uib.de>"]

[[tool.poetry.source]]
name = "uib"
url = "http://pypi.uib.gmbh:8080/simple/"

[tool.poetry.dependencies]
python = ">=3.6, <3.8"
argparse = "^1.4.0"
opsi-dev-tools = "^1.0.6"
python-opsi = "^4.2.0"
pywin32 = { version = "^227", platform = "win32" }
wmi  = { version = "^1.5.1", platform = "win32" }
#mysqlclient = {version = "^1.4.6", platform = "linux" }
mysqlclient = {path = "deps/mysqlclient-1.4.6-cp37-cp37m-win32.whl", platform = "win32"}

[tool.poetry.scripts]
hwaudit = "hwaudit.__main__:main"

[tool.poetry.dev-dependencies]
pytest = "^4.6"
pyinstaller = "^3.6"
pywin32-ctypes = { version = "^0.2.0", platform = "win32" }
pefile = "^2019.4.18"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"